Update GBL start times for practical date legality
Rufflet, Scraggy, Galarian Farfetch'd, and Season 6 Kyurem couldn't
always be obtained the first day of the season due to ranks required.
Maximum number of sets per day is 5 (5 battles each), with reset at
midnight local time. This adjusts start dates for each season's
requirements. If a Pokémon required 3 sessions to be obtainable, the
start date was set to be 3 later and then -1 checked to allow for
players in earlier timezones.