More cleanup
No point stating if a Raid Boss was available during an event, if said Raid Boss was available over the course of 1 month+. It could've been in a number of events, so it's not specific to that event.
Remove end date on Research Breakthrough Scraggy, this went unnoticed for a while. Subsequent GBL encounters are now irrelevant!