Add GBL Landorus encounter
Season 6 ends tomorrow at 1PM PST, so everyone's ranks will be reset back to 1. This encounter is under the assumption that you encountered Landorus through GO Battle League while at Rank 20 in Season 6.

Season 7 starts right when Season 6 ends, so everyone's Rank will be reset back to 1. Season 7 has the same Rank Up structure as Season 6, so it will take at least a week for anyone to reach Rank 20 to encounter a Legendary... but the problem is that Landorus leaves Raids on March 6th (local time). Because you can only play 5 sets per day, this makes it impossible to encounter Landorus before it leaves the encounter pool considering the requirements to Rank Up are too steep with only 6 days to play.