Update Ultra Bonus Event encounters
Ultra Bonus Event 2018 was its own event that spanned 17 days, whereas Ultra Bonus Event 2019 was spread out over the course of 3 events (Johto Journey, Global Challenge, Global Hatches, and A Unova Unveiling). Since that seems to be the new norm (2020 and now 2021 are the same), we can remove 2018 from the event name since it was one of a kind.