Early Combee bugs 🐝
So, apparently when they debuted Combee, there was an issue where the evolution button was visible for Male Combee. Instead of keeping it visible until a proper fix was pushed, Niantic instead decided to modify Combee's gender ratio from 87.5/12.5 to 0/100... so only Female Combee could be encountered in the wild. Male Combee was nowhere to be seen.

When the change happened: https://twitter.com/Kelven91/status/1091151179411791873
When the bug was fixed: https://twitter.com/Kelven91/status/1095819630508208128