Add Egg encounter for male Frillish
So, the official website for Season of GO lists female Frillish as a potential encounter from Adventure Sync 5km Eggs, but after I received one ingame today, it lists male Frillish instead. Not sure if they intended for this, or the website just has inaccurate info, but this is now possible.