Handle early preparations for GO Safari Ball handling (#5)
Updates `PogoType.cs` to account for more specific encounter methods available in-game, in order to account for certain rules in which what Poké Balls are available to use when encountered.

As a result, `data.json` has been significantly modified to reflect these changes, and a majority of logged encounters have had their old `Type` fields remapped to these new IDs.